backend/pyproject.toml
[project]
name = "raindrop-project"
version = "0.1.0"
description = "Add your description here"
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "clickhouse-connect>=0.10.0",
    "fastapi>=0.124.4",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "lark>=1.2.0",
]
